Created LabelFieldSelectorView to reduce repeated code.
The Label, Field, ContextMenu, Toggle Button, Button group for logbooks
and tags was largely the same however they differed in the methods they
called to operate on different lists. The new class provides a generic
view that has the list methods passed to it as arguments. This cuts down
on duplicate code.