Changed when log entry submission occurs. Added setModelTemplate method.
The logbook entry will  be submitted when the dialog's submit button is
pressed. The log entry is created, a new thread is spawned to handle the
network IO of actually sending the entry, and the log entry that was
created is returned from the dialog.

The setModelTemplate takes a LogEntry as an argument and allows the
LogEntryDialog to be initialized to default contents.