Adding check box to menu item in log book and tags context menu.
The context menu's items are now all CustomMenuItems containing
CheckBox's. The CheckBox's text field is set to the log book or tag
name. The context menu's no longer close on a single selection but
instead allow a user to select multiple log books or tags before the
context menu loses the focus.